Estructura de la Aplicación Mejorada

+++ Proyecto/A-History-Fantasy/src/main/java/es/urjc/etsii/dad/Entities/Batalla.java
@@ -6,6 +6,6 @@ import javax.persistence.*;
 public class Batalla {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
-	private int Tipo;
+	private Long id;
+	private long Tipo;
 }
